compress boost/fave notifications more
